Don't force release_status to 'stable'
When using the default metadata (i.e. when the dist has no metadata file),
do not assume that the release_status is 'stable'.  If we assert that
a release is stable, but the version number indicates otherwise, then
CPAN::Meta blows up with this kind of error:

Failed to clean-up 2 metadata. Errors:
'stable' for 'release_status' is invalid for version '0.1_1' (release_status) [Validation: 2]
 at .../Dist/Metadata.pm line 211

By not setting the release_status in the default metadata, CPAN::Meta will
compute it for us, by examining the version number for underscores.

I also added a test case to verify this behavior.